Tweak model.hh and model_nested_test
* Add 'const' qualifiers
* Remove potentially narrowing conversion from short to int

+++ dune/utopia/core/test/model_nested_test.hh
@@ -36,7 +36,7 @@ public:
         // Pass arguments to the base class constructor
         Base(name, parent_model),
         // Store level
-        _level(as_<unsigned short>(this->_cfg["level"]))
+        _level(as_<unsigned int>(this->_cfg["level"]))
     {
         this->_log->info("DoNothingModel initialized. Level: {}", _level);
     }
@@ -76,7 +76,7 @@ public:
         // Pass arguments to the base class constructor
         Base(name, parent_model),
         // Store level
-        _level(as_<unsigned short>(this->_cfg["level"])),
+        _level(as_<unsigned int>(this->_cfg["level"])),
         // Submodel initialization
         lazy("lazy", *this)
     {
@@ -124,7 +124,7 @@ public:
         // Pass arguments to the base class constructor
         Base(name, parent_model),
         // Store level
-        _level(as_<unsigned short>(this->_cfg["level"])),
+        _level(as_<unsigned int>(this->_cfg["level"])),
         // Submodel initialization
         sub_one("one", *this),
         sub_lazy("lazy", *this)
@@ -171,7 +171,7 @@ public:
         // Initialize completely via parent class constructor
         Base(name, parent_model),
         // Store level
-        _level(as_<unsigned short>(this->_cfg["level"])),
+        _level(as_<unsigned int>(this->_cfg["level"])),
         // Submodel initialization
         sub_one("one", *this),
         sub_another("another", *this)
